Rename shpdctl to shepctl
Related to: #63
