+++ scripts/install.sh
@@ -7,7 +7,7 @@ if [ -z "$VER" ]; then
 fi
 
 INSTALL_DIR=${INSTALL_DIR:-"/usr/local/bin"}
-URL="https://github.com/LunaticFringers/shepherd/releases/download/v$VER/shpdctl-$VER.tar.gz"
+URL="https://github.com/LunaticFringers/shepherd/releases/download/v$VER/shepctl-$VER.tar.gz"
 TEMP_DIR=$(mktemp -d)
 
 if [ "$(id -u)" -ne 0 ]; then
@@ -15,21 +15,21 @@ if [ "$(id -u)" -ne 0 ]; then
   exit 1
 fi
 
-echo "Installing shpdctl version $VER..."
+echo "Installing shepctl version $VER..."
 echo "Downloading from URL $URL"
-curl -fsSL "$URL" -o "$TEMP_DIR/shpdctl.tar.gz"
+curl -fsSL "$URL" -o "$TEMP_DIR/shepctl.tar.gz"
 
 echo "Extracting the package..."
-tar -xzvf "$TEMP_DIR/shpdctl.tar.gz" -C "$TEMP_DIR"
+tar -xzvf "$TEMP_DIR/shepctl.tar.gz" -C "$TEMP_DIR"
 
-echo "Installing shpdctl binary to $INSTALL_DIR..."
-mv "$TEMP_DIR/shpdctl" "$INSTALL_DIR/shpdctl"
+echo "Installing shepctl binary to $INSTALL_DIR..."
+mv "$TEMP_DIR/shepctl" "$INSTALL_DIR/shepctl"
 
-chmod +x "$INSTALL_DIR/shpdctl"
+chmod +x "$INSTALL_DIR/shepctl"
 rm -rf "$TEMP_DIR"
 
-if command -v shpdctl > /dev/null; then
-  echo "shpdctl installed successfully! You can run it using the 'shpdctl' command."
+if command -v shepctl > /dev/null; then
+  echo "shepctl installed successfully! You can run it using the 'shepctl' command."
 else
   echo "Installation failed."
   exit 1
